+++ src/main/scala/com/ucesys/sparkscope/metrics/StageMetrics.scala
@@ -3,7 +3,7 @@ package com.ucesys.sparkscope.metrics
 import com.ucesys.sparkscope.common.StageContext
 import com.ucesys.sparkscope.data.{DataColumn, DataTable}
 
-case class StageMetrics(stageTimeline: DataTable, clusterCPUCapacity: DataTable) {
+case class StageMetrics(stageTimeline: DataTable, numberOfTasks: DataColumn) {
     override def toString: String = {
         Seq(
             s"\nStage metrics:",
@@ -13,8 +13,7 @@ case class StageMetrics(stageTimeline: DataTable, clusterCPUCapacity: DataTable)
 }
 
 object StageMetrics {
-    def apply(stages: Seq[StageContext], allTimestamps: Seq[String], clusterCPUCapacity: DataTable): StageMetrics = {
-        // Stages
+    def apply(stages: Seq[StageContext], allTimestamps: Seq[String]): StageMetrics = {
         val stageColumns: Seq[DataColumn] = stages.map { stage =>
             val colVals: Seq[String] = allTimestamps.map(_.toLong).map { ts =>
                 if (ts > stage.getTimelineStart && ts < stage.getTimelineEnd) {
@@ -28,21 +27,20 @@ object StageMetrics {
             DataColumn(stage.stageId, colVals)
         }
 
-        val clusterCPUCapacityAligned =  {
-            val existingTimestamps = clusterCPUCapacity.select("t").values
-            val newTimestamps = allTimestamps.filterNot(existingTimestamps.contains)
-            DataTable(
-                "heapUsed",
-                Seq(
-                    DataColumn("t", clusterCPUCapacity.select("t").values ++ newTimestamps),
-                    DataColumn("totalCores", clusterCPUCapacity.select("totalCores").values ++ Seq.fill(newTimestamps.length)("null")),
-                )
-            ).sortBy("t")
+        val numberOfTasksVals: Seq[String] = allTimestamps.map(_.toLong).map { ts =>
+            val numberOfTasksSum: Long = stages.map { stage =>
+                if (ts >= stage.getTimelineStart && ts <= stage.getTimelineEnd) {
+                    stage.numberOfTasks
+                } else {
+                    0L
+                }
+            }.sum
+            numberOfTasksSum.toString
         }
 
         StageMetrics(
             stageTimeline = DataTable("stages", Seq(DataColumn("t", allTimestamps)) ++ stageColumns),
-            clusterCPUCapacity = clusterCPUCapacityAligned
+            numberOfTasks = DataColumn("numberOfTasks", numberOfTasksVals)
         )
     }
 }